cluster: extract receive/dispatch path into sync_conn_read.go
Fifth pure code-motion increment of the #5661 sync_conn.go split. Moves
the inbound framing/read loop and the message dispatch switch into a
cohesive sibling file, leaving sync_conn.go as the connection-lifecycle
core (dial, accept, install, start/stop, disconnect).

Moved verbatim (no rename, no logic/timing change):
  receiveLoop, handleMessage.

Same package (pkg/cluster); imports recomputed per file. Behavior is
byte-identical. sync_conn.go drops from 2228 to 617 lines.

go build ./... passes.
